+++ pedantic/decorators/cls_deco_frozen_dataclass.py
@@ -1,23 +1,44 @@
 from dataclasses import dataclass, asdict, fields, is_dataclass
-from typing import Type, TypeVar, Any
+from typing import Type, TypeVar, Any, Callable
 
 from pedantic.type_checking_logic.check_types import assert_value_matches_type
 
 T = TypeVar('T')
 
 
-def frozen_dataclass(cls: Type[T]) -> Type[T]:
+def frozen_type_safe_dataclass(cls: Type[T]) -> Type[T]:
+    """ Shortcut for @frozen_dataclass(type_safe=True) """
+
+    return frozen_dataclass(type_safe=True)(cls)
+
+
+def frozen_dataclass(cls: Type[T] = None, type_safe: bool = False, order: bool = False) -> Type[T]:
     """
-        Makes the decorated class immutable by adding the [@dataclass(frozen=True)]
-        decorator and add a useful [copy_with()] instance method to this class.
+        Makes the decorated class immutable and a dataclass by adding the [@dataclass(frozen=True)]
+        decorator. Also adds useful copy_with() and validate_types() instance methods to this class (see below).
+
+        If [type_safe] is True, a type check is performed for each field after the __post_init__ method was called
+        which itself s directly called after the __init__ constructor.
+        Note this have a negative impact on the performance. It's recommend to use this for debugging and testing only.
 
         In a nutshell, the followings methods will be added to the decorated class automatically:
-        - __init__()
-        - __eq__()
-        - __hash__()
-        - __repr__()
-        - copy_with()
-        - validate_types()
+        - __init__() gives you a simple constructor like "Foo(a=6, b='hi', c=True)"
+        - __eq__() lets you compare objects easily with "a == b"
+        - __hash__() is also needed for instance comparison
+        - __repr__() gives you a nice output when you call "print(foo)"
+        - copy_with() allows you to quickly create new similar frozen instances. Use this instead of setters.
+        - validate_types() allows you to validate the types of the dataclass.
+                           This is called automatically when [type_safe] is True.
+
+        If the [order] parameter is True (default is False), the following comparison methods
+        will be added additionally:
+        - __lt__() lets you compare instance like "a < b"
+        - __le__() lets you compare instance like "a <= b"
+        - __gt__() lets you compare instance like "a > b"
+        - __ge__() lets you compare instance like "a >= b"
+
+        These compare the class as if it were a tuple of its fields, in order.
+        Both instances in the comparison must be of the identical type.
 
         Example:
 
@@ -41,36 +62,61 @@ def frozen_dataclass(cls: Type[T]) -> Type[T]:
         Foo(a=676676, b='new', c=False)
     """
 
-    def copy_with(self, **kwargs: Any) -> T:
-        """
-            Creates a new immutable instance that by copying all fields of this instance replaced by the new values.
-        """
+    def decorator(cls_: Type[T]) -> Type[T]:
+        def copy_with(self, **kwargs: Any) -> T:
+            """
+                Creates a new immutable instance that by copying all fields of this instance replaced by the new values.
+            """
 
-        return cls(**{**asdict(self), **kwargs})  # python >= 3.9.0: cls(**(asdict(self) | kwargs))
+            return cls_(**{**asdict(self), **kwargs})  # python >= 3.9.0: cls(**(asdict(self) | kwargs))
 
-    def validate_types(self) -> None:
-        """
-            Checks that all instance variable have the correct type.
-            Raises an Exception if at least one type is incorrect.
-        """
+        def validate_types(self) -> None:
+            """
+                Checks that all instance variable have the correct type.
+                Raises a [PedanticTypeCheckException] if at least one type is incorrect.
+            """
 
-        props = fields(cls)
+            props = fields(cls_)
 
-        for field in props:
-            assert_value_matches_type(
-                value=getattr(self, field.name),
-                type_=field.type,
-                err=f'In dataclass "{cls.__name__}" in field "{field.name}": ',
-                type_vars={},
-            )
+            for field in props:
+                assert_value_matches_type(
+                    value=getattr(self, field.name),
+                    type_=field.type,
+                    err=f'In dataclass "{cls_.__name__}" in field "{field.name}": ',
+                    type_vars={},
+                )
 
-    if is_dataclass(obj=cls):
-        raise AssertionError(f'Dataclass "{cls}" cannot be decorated with '
-                             f'"@frozen_dataclass" because it already is a dataclass.')
+        def _post_init(f: Callable[[T], None]) -> Callable[[T], None]:
+            def wrapper(self) -> None:
+                f(self)
+                self.validate_types()
 
-    setattr(cls, 'copy_with', copy_with)
-    setattr(cls, 'validate_types', validate_types)
-    return dataclass(frozen=True)(cls)
+            return wrapper
+
+        if is_dataclass(obj=cls_):
+            raise AssertionError(f'Dataclass "{cls_}" cannot be decorated with '
+                                 f'"@frozen_dataclass" because it already is a dataclass.')
+
+        setattr(cls_, copy_with.__name__, copy_with)
+        setattr(cls_, validate_types.__name__, validate_types)
+
+        if type_safe:
+            post_init = '__post_init__'
+
+            if hasattr(cls_, post_init):
+                new_post_init = _post_init(f=getattr(cls_, post_init))
+            else:
+                def new_post_init(self):
+                    self.validate_types()
+
+            setattr(cls_, post_init, new_post_init)
+
+        return dataclass(frozen=True, order=order)(cls_)
+
+    if cls is None:
+        return decorator
+
+    return decorator(cls_=cls)
 
 
 if __name__ == '__main__':